Refactor to use image library
