+++ package.json
@@ -24,6 +24,7 @@
     "agenda": "^0.9.0",
     "all-the-cities": "^2.0.0",
     "async": "^2.3.0",
+    "aws-sdk": "2.54.0",
     "bluebird": "^3.5.0",
     "catbox-memory": "2.0.4",
     "cities": "1.1.2",
